Bug fixes, design, and CR feedback
- Fixed filter removing student events
- Moved filter header to table view header so pull to refresh is at the
top
- Moved LoadingCell to its own file and made it public
